+++ test/unit/rx-query.test.ts
@@ -393,7 +393,7 @@ config.parallel('rx-query.test.js', () => {
             // it is assumed that this query can never handled by the QueryChangeDetector
             const query = col.find().sort('-passportId').limit(1);
 
-            const fired: any [] = [];
+            const fired: any[] = [];
             const sub1 = query.$.subscribe(res => {
                 fired.push(res);
             });
@@ -1168,6 +1168,16 @@ config.parallel('rx-query.test.js', () => {
                             $gt: null
                         }
                     }
+                }, {
+                    user_id: {
+                        $eq: '6'
+                    }
+                },
+                {
+                    created_at: {
+                        $gt: null
+                    }
+                }
                 ]
             };
             /* eslint-enable */
@@ -1349,9 +1359,72 @@ config.parallel('rx-query.test.js', () => {
 
             db.destroy();
         });
-    });
+        /**
+         * via gitter at 11 November 2019 10:10
+         */
+        it('gitter: query with regex does not return correct results', async () => {
+            // create a schema
+            const mySchema = {
+                version: 0,
+                type: 'object',
+                properties: {
+                    passportId: {
+                        type: 'string',
+                        primary: true
+                    },
+                    firstName: {
+                        type: 'string'
+                    },
+                    lastName: {
+                        type: ['string', 'null']
+                    }
+                }
+            };
+            const db = await RxDB.create({
+                name: util.randomCouchString(10),
+                adapter: 'memory',
+                queryChangeDetection: true,
+                ignoreDuplicate: true
+            });
 
-    describe('e', () => {
-        // it('e', () => process.exit());
+            // create a collection
+            const collection = await db.collection({
+                name: 'mycollection',
+                schema: mySchema
+            });
+
+            // insert documents
+            await collection.bulkInsert([
+                {
+                    passportId: 'doc1',
+                    firstName: 'John',
+                    lastName: 'Doe'
+                }, {
+                    passportId: 'doc2',
+                    firstName: 'Martin',
+                    lastName: 'Smith'
+                }
+            ]);
+            const allDocs = await collection.find().exec();
+            assert.strictEqual(allDocs.length, 2);
+
+            // test 1 with RegExp object
+            const regexp = new RegExp('^Doe$', 'i');
+            const result1 = await collection.find({ lastName: { $regex: regexp } }).exec();
+
+            // test 2 with regex string
+            const result2 = await collection.find({ lastName: { $regex: '^Doe$' } }).exec();
+
+
+            // both results should only have the doc1
+            assert.strictEqual(result1.length, 1);
+            assert.strictEqual(result1[0].passportId, 'doc1');
+            assert.deepStrictEqual(
+                result1.map(d => d.toJSON()),
+                result2.map(d => d.toJSON())
+            );
+
+            db.destroy();
+        });
     });
 });